requirements.txt
google-cloud-aiplatform
python-dotenv
jsonschema
gTTS
pydub
ffmpeg
pyttsx3
pyobjc
